Fix: Use real-*-command
Evil sets `this-command` to `previous-line` and `next-line` for a lot of
scrolling commands (e.g. `evil-scroll-up` and `evil-scroll-down`). This
causes `restore-point` to not save the point when scrolling. This commit
fixes this by using `real-this-command` and `real-last-command` instead
of `this-command` and `last-command`.

Also, ensure that the `rp/restore-point-position` function does not throw
an error when `rp/point-ring` is nil.